+++ test/libsolidity/syntaxTests/types/address/uint_to_payable_address.sol
@@ -1,5 +1,5 @@
 contract C {
     function f(uint x) public pure returns (address payable) {
-        return address(x);
+        return address(uint160(x));
     }
 }
